Changes: - fixing bug on server - adding sound onMouse enter on home Button and on Click

+++ server/Helper/userCollection.js
@@ -9,7 +9,7 @@ function createUserCollection(db){
                  description: "must be a string and is required"
               },
               score: {
-                 bsonType : "double",
+                 bsonType : "int",
               }
            }
         } }
